fixed scaling and added sections to a disk for a future scaling change

+++ DataVisualization/Disk.h
@@ -3,8 +3,8 @@
 
 class Disk: public ShapeManager {
 private:
-	int stride;
-	int padding = 0;
+	double stride;
+	double padding = 0;
 public:
 	Disk(GLFWwindow* window, int bindWidth, int bindHeight, int posX, int posY);
 	~Disk();
